requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
jwt==1.3.1
pycparser==2.22
requests==2.32.3
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0